Added a second Event for the 'door warning annunciator'
and switched from bytewise to bitwise changes (unsigned byte also
changed to unsigned word.)
Attention: Works only with modified RASDuke.cab in the panels folder of
the RealAir Duke. This is caused by the lack of the new added
'L:CabinDoor_Annunciator' which doesn't exist in the original version
from Realair!